backend/requirements.txt
Flask==2.3.2
gunicorn==21.2.0
flask-cors==4.0.0
psycopg2
python-dotenv==1.0.0 # Jeśli używasz .env lokalnie
cloudinary # Dodaj tę linię
Werkzeug==2.3.7